Ephemereal queries shouldn't be using .execute
These cache prepared statements, which is desirable for most
queries (just plug in new values), but slowly starts to exhaust
the statement cache. We can run out within a day.

Unpreparing the statement is additional overhead for ephemereal
queries, so we'll just use an uncached substition instead.

This does put some additional onus to refactor the search (which
works right now, but has some inefficiency and simplicity issues).